+++ app/Http/Controllers/rentAfriend/rentAfriendOrderController.php
@@ -134,7 +134,7 @@ class rentAfriendOrderController extends Controller
             $charge = $stripe->charges->create([
                 "card" => $token["id"],
                 "currency" => "USD",
-                "amount" => ($rentAfriend_order->sub_total + $rentAfriend_order->tax) * 100,
+                "amount" => (int) (($rentAfriend_order->sub_total + $rentAfriend_order->tax) * 100),
                 "description" => "Pay rent a friend Order",
             ]);
